services/frontend/requirements.txt
nicegui
httpx
python-dotenv